petrescue/requirements.txt
Pillow>=10.0.0
imagehash>=4.3
# Optional (better matching):
# opencv-python
requests>=2.0
tensorflow>=2.12.0
scikit-learn>=1.2

django>=4.0.0
pymysql>=1.0.0
twilio>=8.0.0